git
Add cors package to backend/server.js to reolve cross-origin error. In front end we
implement hooks useState and useEffect in our Home component. We pass down the workout property to our new Workout Details component which will render and individual workout. In the Home componenent we map through the workouts and import the WorkoutDetails component to render each workout to the web page